// app/dashboard/(ecommerce)/_types/Product/ProductTagModel.d.ts
/**
 * Represents a product tag.
 */
export interface ProductTagModel {
    /**
     * The ID of the product tag.
     */
    id: number;
  
    /**
     * The name of the product tag.
     */
    name: string;
  
    /**
     * The number of products associated with this tag.
     */
    products: number;
  }